server/requirements.txt
flask
flask-cors
pyjwt
python-dotenv
supabase
gunicorn